Refactor asterisc/cannon provider using utils
