+++ op-challenger/game/fault/trace/cannon/provider.go
@@ -9,11 +9,10 @@ import (
 	"math"
 	"os"
 	"path/filepath"
-	"strconv"
 
 	"github.com/ethereum-optimism/optimism/op-challenger/config"
+	"github.com/ethereum-optimism/optimism/op-challenger/game/fault/trace/utils"
 	"github.com/ethereum-optimism/optimism/op-challenger/game/fault/types"
-	preimage "github.com/ethereum-optimism/optimism/op-preimage"
 	"github.com/ethereum-optimism/optimism/op-program/host/kvstore"
 	"github.com/ethereum-optimism/optimism/op-service/ioutil"
 	"github.com/ethereum/go-ethereum/common"
@@ -23,50 +22,31 @@ import (
 	"github.com/ethereum-optimism/optimism/cannon/mipsevm"
 )
 
-const (
-	proofsDir      = "proofs"
-	diskStateCache = "state.json.gz"
-)
-
-type ProofData struct {
-	ClaimValue   common.Hash   `json:"post"`
-	StateData    hexutil.Bytes `json:"state-data"`
-	ProofData    hexutil.Bytes `json:"proof-data"`
-	OracleKey    hexutil.Bytes `json:"oracle-key,omitempty"`
-	OracleValue  hexutil.Bytes `json:"oracle-value,omitempty"`
-	OracleOffset uint32        `json:"oracle-offset,omitempty"`
-}
-
 type CannonMetricer interface {
 	RecordCannonExecutionTime(t float64)
-}
-
-type ProofGenerator interface {
-	// GenerateProof executes FPVM binary to generate a proof at the specified trace index in dataDir.
-	GenerateProof(ctx context.Context, dataDir string, proofAt uint64) error
 }
 
 type CannonTraceProvider struct {
 	logger         log.Logger
 	dir            string
 	prestate       string
-	generator      ProofGenerator
+	generator      utils.ProofGenerator
 	gameDepth      types.Depth
-	preimageLoader *PreimageLoader
+	preimageLoader *utils.PreimageLoader
 
 	// lastStep stores the last step in the actual trace if known. 0 indicates unknown.
 	// Cached as an optimisation to avoid repeatedly attempting to execute beyond the end of the trace.
 	lastStep uint64
 }
 
-func NewTraceProvider(logger log.Logger, m CannonMetricer, cfg *config.Config, localInputs LocalGameInputs, dir string, gameDepth types.Depth) *CannonTraceProvider {
+func NewTraceProvider(logger log.Logger, m CannonMetricer, cfg *config.Config, localInputs utils.LocalGameInputs, dir string, gameDepth types.Depth) *CannonTraceProvider {
 	return &CannonTraceProvider{
 		logger:         logger,
 		dir:            dir,
 		prestate:       cfg.CannonAbsolutePreState,
 		generator:      NewExecutor(logger, m, cfg, localInputs),
 		gameDepth:      gameDepth,
-		preimageLoader: NewPreimageLoader(kvstore.NewDiskKV(PreimageDir(dir)).Get),
+		preimageLoader: utils.NewPreimageLoader(kvstore.NewDiskKV(utils.PreimageDir(dir)).Get),
 	}
 }
 
@@ -137,10 +117,10 @@ func (p *CannonTraceProvider) AbsolutePreStateCommitment(_ context.Context) (com
 
 // loadProof will attempt to load or generate the proof data at the specified index
 // If the requested index is beyond the end of the actual trace it is extended with no-op instructions.
-func (p *CannonTraceProvider) loadProof(ctx context.Context, i uint64) (*ProofData, error) {
+func (p *CannonTraceProvider) loadProof(ctx context.Context, i uint64) (*utils.ProofData, error) {
 	// Attempt to read the last step from disk cache
 	if p.lastStep == 0 {
-		step, err := ReadLastStep(p.dir)
+		step, err := utils.ReadLastStep(p.dir)
 		if err != nil {
 			p.logger.Warn("Failed to read last step from disk cache", "err", err)
 		} else {
@@ -151,7 +131,7 @@ func (p *CannonTraceProvider) loadProof(ctx context.Context, i uint64) (*ProofDa
 	if p.lastStep != 0 && i > p.lastStep {
 		i = p.lastStep
 	}
-	path := filepath.Join(p.dir, proofsDir, fmt.Sprintf("%d.json.gz", i))
+	path := filepath.Join(p.dir, utils.ProofsDir, fmt.Sprintf("%d.json.gz", i))
 	file, err := ioutil.OpenDecompressed(path)
 	if errors.Is(err, os.ErrNotExist) {
 		if err := p.generator.GenerateProof(ctx, p.dir, i); err != nil {
@@ -177,7 +157,7 @@ func (p *CannonTraceProvider) loadProof(ctx context.Context, i uint64) (*ProofDa
 				if err != nil {
 					return nil, fmt.Errorf("cannot hash witness: %w", err)
 				}
-				proof := &ProofData{
+				proof := &utils.ProofData{
 					ClaimValue:   witnessHash,
 					StateData:    hexutil.Bytes(witness),
 					ProofData:    []byte{},
@@ -185,7 +165,7 @@ func (p *CannonTraceProvider) loadProof(ctx context.Context, i uint64) (*ProofDa
 					OracleValue:  nil,
 					OracleOffset: 0,
 				}
-				if err := WriteLastStep(p.dir, proof, p.lastStep); err != nil {
+				if err := utils.WriteLastStep(p.dir, proof, p.lastStep); err != nil {
 					p.logger.Warn("Failed to write last step to disk cache", "step", p.lastStep)
 				}
 				return proof, nil
@@ -198,7 +178,7 @@ func (p *CannonTraceProvider) loadProof(ctx context.Context, i uint64) (*ProofDa
 		return nil, fmt.Errorf("cannot open proof file (%v): %w", path, err)
 	}
 	defer file.Close()
-	var proof ProofData
+	var proof utils.ProofData
 	err = json.NewDecoder(file).Decode(&proof)
 	if err != nil {
 		return nil, fmt.Errorf("failed to read proof (%v): %w", path, err)
@@ -207,43 +187,11 @@ func (p *CannonTraceProvider) loadProof(ctx context.Context, i uint64) (*ProofDa
 }
 
 func (c *CannonTraceProvider) finalState() (*mipsevm.State, error) {
-	state, err := parseState(filepath.Join(c.dir, finalState))
+	state, err := parseState(filepath.Join(c.dir, utils.FinalState))
 	if err != nil {
 		return nil, fmt.Errorf("cannot read final state: %w", err)
 	}
 	return state, nil
-}
-
-type diskStateCacheObj struct {
-	Step uint64 `json:"step"`
-}
-
-// ReadLastStep reads the tracked last step from disk.
-func ReadLastStep(dir string) (uint64, error) {
-	state := diskStateCacheObj{}
-	file, err := ioutil.OpenDecompressed(filepath.Join(dir, diskStateCache))
-	if err != nil {
-		return 0, err
-	}
-	defer file.Close()
-	err = json.NewDecoder(file).Decode(&state)
-	if err != nil {
-		return 0, err
-	}
-	return state.Step, nil
-}
-
-// WriteLastStep writes the last step and proof to disk as a persistent cache.
-func WriteLastStep(dir string, proof *ProofData, step uint64) error {
-	state := diskStateCacheObj{Step: step}
-	lastStepFile := filepath.Join(dir, diskStateCache)
-	if err := ioutil.WriteCompressedJson(lastStepFile, state); err != nil {
-		return fmt.Errorf("failed to write last step to %v: %w", lastStepFile, err)
-	}
-	if err := ioutil.WriteCompressedJson(filepath.Join(dir, proofsDir, fmt.Sprintf("%d.json.gz", step)), proof); err != nil {
-		return fmt.Errorf("failed to write proof: %w", err)
-	}
-	return nil
 }
 
 // CannonTraceProviderForTest is a CannonTraceProvider that can find the step referencing the preimage read
@@ -252,49 +200,19 @@ type CannonTraceProviderForTest struct {
 	*CannonTraceProvider
 }
 
-type preimageOpts []string
-
-type PreimageOpt func() preimageOpts
-
-func PreimageLoad(key preimage.Key, offset uint32) PreimageOpt {
-	return func() preimageOpts {
-		return []string{"--stop-at-preimage", fmt.Sprintf("%v@%v", common.Hash(key.PreimageKey()).Hex(), offset)}
-	}
-}
-
-func FirstPreimageLoadOfType(preimageType string) PreimageOpt {
-	return func() preimageOpts {
-		return []string{"--stop-at-preimage-type", preimageType}
-	}
-}
-
-func FirstKeccakPreimageLoad() PreimageOpt {
-	return FirstPreimageLoadOfType("keccak")
-}
-
-func FirstPrecompilePreimageLoad() PreimageOpt {
-	return FirstPreimageLoadOfType("precompile")
-}
-
-func PreimageLargerThan(size int) PreimageOpt {
-	return func() preimageOpts {
-		return []string{"--stop-at-preimage-larger-than", strconv.Itoa(size)}
-	}
-}
-
-func NewTraceProviderForTest(logger log.Logger, m CannonMetricer, cfg *config.Config, localInputs LocalGameInputs, dir string, gameDepth types.Depth) *CannonTraceProviderForTest {
+func NewTraceProviderForTest(logger log.Logger, m CannonMetricer, cfg *config.Config, localInputs utils.LocalGameInputs, dir string, gameDepth types.Depth) *CannonTraceProviderForTest {
 	p := &CannonTraceProvider{
 		logger:         logger,
 		dir:            dir,
 		prestate:       cfg.CannonAbsolutePreState,
 		generator:      NewExecutor(logger, m, cfg, localInputs),
 		gameDepth:      gameDepth,
-		preimageLoader: NewPreimageLoader(kvstore.NewDiskKV(PreimageDir(dir)).Get),
+		preimageLoader: utils.NewPreimageLoader(kvstore.NewDiskKV(utils.PreimageDir(dir)).Get),
 	}
 	return &CannonTraceProviderForTest{p}
 }
 
-func (p *CannonTraceProviderForTest) FindStep(ctx context.Context, start uint64, preimage PreimageOpt) (uint64, error) {
+func (p *CannonTraceProviderForTest) FindStep(ctx context.Context, start uint64, preimage utils.PreimageOpt) (uint64, error) {
 	// Run cannon to find the step that meets the preimage conditions
 	if err := p.generator.(*Executor).generateProof(ctx, p.dir, start, math.MaxUint64, preimage()...); err != nil {
 		return 0, fmt.Errorf("generate cannon trace (until preimage read): %w", err)
