+++ op-challenger/game/fault/trace/utils/local_test.go
@@ -1,4 +1,4 @@
-package cannon
+package utils
 
 import (
 	"context"
