+++ op-challenger/game/fault/trace/utils/preimage.go
@@ -1,4 +1,4 @@
-package cannon
+package utils
 
 import (
 	"bytes"
